rename pyspec pkg to eth2spec
