+++ test_libs/pyspec/eth2spec/utils/bls_stub.py
