+++ test_libs/pyspec/eth2spec/utils/merkle_minimal.py
