+++ test_libs/pyspec/eth2spec/utils/minimal_ssz.py
